requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==4.2.6
django-allauth==0.57.2
django-cors-headers==4.2.0
django-ipware==6.0.4
django-rest-authemail==2.1.7
djangorestframework==3.14.0
idna==3.6
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
python-ipware==2.0.2
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.4.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.2.1
